requirements.txt
sqlparse==0.5.0
psycopg2-binary==2.9.9
streamlit==1.35.0
anthropic==0.25.0
python-dotenv==1.0.1
